--- src/main/java/synopsis/graphql/excpetion/CustomGraphqlExceptionResolver.java
@@ -1,45 +1,0 @@
-package synopsis.graphql.excpetion;
-
-import graphql.ErrorType;
-import graphql.GraphQLError;
-import graphql.GraphqlErrorBuilder;
-import graphql.schema.DataFetchingEnvironment;
-import org.jetbrains.annotations.NotNull;
-import org.springframework.graphql.execution.DataFetcherExceptionResolverAdapter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class CustomGraphqlExceptionResolver extends DataFetcherExceptionResolverAdapter {
-
-    @Override
-    protected GraphQLError resolveToSingleError(@NotNull Throwable ex, @NotNull DataFetchingEnvironment env) {
-        if (ex instanceof ServerRequestException) {
-            return GraphqlErrorBuilder.newError()
-                    .errorType(ErrorType.ExecutionAborted)
-                    .message(ex.getMessage())
-                    .path(env.getExecutionStepInfo().getPath())
-                    .location(env.getField().getSourceLocation())
-                .build();
-
-        } else if (ex instanceof JsonToObjectException) {
-            return GraphqlErrorBuilder.newError()
-                    .errorType(ErrorType.ValidationError)
-                    .message(ex.getMessage())
-                    .path(env.getExecutionStepInfo().getPath())
-                    .location(env.getField().getSourceLocation())
-                .build();
-
-        } else if (ex instanceof CustomJsonPropertyException) {
-            return GraphqlErrorBuilder.newError()
-                    .errorType(ErrorType.ValidationError)
-                    .message(ex.getMessage())
-                    .path(env.getExecutionStepInfo().getPath())
-                    .location(env.getField().getSourceLocation())
-                .build();
-        }
-
-        else {
-            return null;
-        }
-    }
-}
